requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# GitHub API
PyGithub==2.1.1

# Google Gemini
google-generativeai==0.3.2

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
